DATAES-372 - Support highlighting via annotation
Refactor CriteriaQuery conversion.
Add implementation highlight support for part query.
