+++ src/test/java/org/springframework/data/elasticsearch/repositories/highlight/FileRepoTests.java
@@ -1,0 +1,7 @@
+package org.springframework.data.elasticsearch.repositories.highlight;
+
+/**
+ * @author zzt
+ */
+public class FileRepoTests {
+}
